mode leds to own file
